brandme-core/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==1.10.15
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2
uuid==1.30
